% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkewedFDA.R
\name{depan}
\alias{depan}
\title{Epanechnikov Kernel function with band-width 1;
for a bandwidth of h, the input should be scaled by h}
\usage{
depan(x)
}
\arguments{
\item{x}{point at which kernel will be evaluated}
}
\value{
a scalar of kernel value
}
\description{
Epanechnikov Kernel function with band-width 1;
for a bandwidth of h, the input should be scaled by h
}
\examples{
depan(0.3)

}
