% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkewedFDA.R
\name{hmma_pvalue}
\alias{hmma_pvalue}
\title{Asymmetric test using Holgersson's statistic}
\usage{
hmma_pvalue(x, boot_ss = 100)
}
\arguments{
\item{x}{a numeric vector of data}

\item{boot_ss}{a scalar represents the bootstrap sample size}
}
\value{
It returns a scalar represents the calculated p-value
}
\description{
It calculates p-value using the Holgersson's modified test statistic to 
test the asymmetry of a given data from bootstrap samples.
}
\examples{
hmma_pvalue(x=rsn(30,dp=cp2dp(c(0,1,0.05),family="SN")),boot_ss=100)

}
